refactor db functions
