+++ app/actions.tsx
@@ -3,7 +3,7 @@
 import { generateId } from "ai";
 import { createAI, getAIState } from "ai/rsc";
 
-import { saveAIStateToStore } from "@/lib/db";
+import { aiState } from "@/lib/db";
 import { confirmPurchase } from "@/llm/actions/confirm-purchase";
 import { continueConversation } from "@/llm/actions/continue-conversation";
 import { checkEnrollment } from "@/llm/actions/newsletter";
@@ -31,7 +31,7 @@ export const AI = (p: Props) => {
 
       if (done) {
         const user = await getUser();
-        await saveAIStateToStore({
+        await aiState.save({
           conversationID,
           messages: state,
           userID: user.sub,
