+++ llm/actions/history.tsx
@@ -1,7 +1,7 @@
 "use server";
 
-import { getAIStateFromStore } from "@/lib/db";
+import { aiState } from "@/lib/db";
 
 export const getHistoryFromStore = async (id: string) => {
-  return await getAIStateFromStore({ conversationID: id });
+  return await aiState.get({ conversationID: id });
 };
