+++ llm/components/positions.tsx
@@ -2,13 +2,13 @@
 
 import { useActions, useUIState } from "ai/rsc";
 
-import { Position } from "@/lib/db";
+import { transactions } from "@/lib/db";
 import { cn } from "@/lib/utils";
 
 import { ClientMessage } from "../types";
 import WarningWrapper from "./warning-wrapper";
 
-export function Positions({ positions, readOnly = false }: { positions: Position[]; readOnly?: boolean }) {
+export function Positions({ positions, readOnly = false }: { positions: transactions.Position[]; readOnly?: boolean }) {
   const [, setMessages] = useUIState();
   const { continueConversation } = useActions();
